WIP: get CouchDB to use a modern libmozjs.
This is still broken, because (i) libmozjs appears to have incomplete pkg-config
metadata (it leaves off -lnspr4 -lplc4 -lplds4), and (ii) CouchDB v1.6.1's
autoconf script doesn't seem to realize that libmozjs v24 exists.

This will eventually be useful, but it'll take additional work and may prove to
be unstable under CouchDB anyway. When I have more time, I might try rolling
back to libmozjs v17 and see if that works.

The goal of all of this would be to improve performance; libmozjs-185 is ancient
(2011) and is probably dramatically slower (not to mention less secure) than
what's in Firefox in 2015.